fix(folder-structure - folder structure fixed according to page)
