+++ packages/storefront/components/home/imageSlider/singleSlide.component.tsx
