+++ packages/storefront/components/product/product-image-slider.tsx
