--- packages/storefront/pages/dealProducts.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import DealProducts from '../components/deal/products'
-
-const dealProducts = () => {
-    return (
-        <div>
-            <DealProducts></DealProducts>
-        </div>
-    );
-};
-
-export default dealProducts;
