+++ packages/storefront/pages/home.tsx
@@ -1,52 +1,8 @@
 import type { NextPage } from "next";
-import TrendingProducts from "@/components/trend/index";
-import WeekDeals from "@/components/weekDeals";
-import HomeShipping from "@/components/homeShipping";
-import ImageSlider from "@/components/imageSlider";
-import FeaturedProducts from "@/components/featuredProducts";
-import BannerPage from "@/components/global/bannerComponent";
-import BestSell from "@/components/bestSell";
-import Blog from "@/components/blog/blog";
-import HomefullBanner from "@/components/global/bannerComponent/homeFullBanner";
-import { useDispatch } from "react-redux";
-import { storeProducts } from "toolkit/ProductsSlice";
-import { useEffect } from "react";
-import productData from "../allData/product-data.json";
+import HomeComponent from "@/components/home";
 
 const Home: NextPage = () => {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(storeProducts(productData.products));
-  }, []);
-
-  return (
-    <>
-      <ImageSlider />
-      <HomeShipping />
-      <div className="mb-4 md:mb-10">
-        <TrendingProducts />
-      </div>
-      <div className="mb-4 md:mb-10">
-        <BannerPage />
-      </div>
-      <div className="mb-4 md:mb-10">
-        <WeekDeals />
-      </div>
-      <div className="mb-4 md:mb-10">
-        <HomefullBanner />
-      </div>
-      <div className="mb-4 md:mb-10">
-        <BestSell />
-      </div>
-      <div className="mb-5 md:mb-10">
-        <Blog />
-      </div>
-      <div className="mb-4 md:mb-10">
-        <FeaturedProducts />
-      </div>
-    </>
-  );
+  return <HomeComponent />;
 };
 
 export default Home;
