+++ packages/storefront/pages/product/[pid].tsx
@@ -1,13 +1,12 @@
 import type { NextPage } from "next";
-import ProductDetailsComponent from "@/components/productDetails/productDetails";
+import ProductDetailsComponent from "@/components/product";
 
 const ProductDetails: NextPage = () => {
-	
-		return (
-			<>
-				<ProductDetailsComponent></ProductDetailsComponent>
-			</>
-		);
-	}
+  return (
+    <>
+      <ProductDetailsComponent></ProductDetailsComponent>
+    </>
+  );
+};
 
 export default ProductDetails;
